requirements.txt
requests==2.25.1
matplotlib==3.4.2
pandas==1.2.4
numpy==1.20.3
streamlit==0.82.0
beautifulsoup4==4.9.3
Pillow==8.3.1
